Add files via upload
A company has developed a new web page in order to try to increase the number of users who "convert," meaning the number of users who decide to pay for the company's product. The goal is to help the company understand if they should implement this new page, keep the old page, or perhaps run the experiment longer to make their decision. A/B testing technique is to be used to achieve the goal.